ib_client_service/requirements.txt
fastapi
uvicorn[standard]
ib_async
python-dotenv
pandas
